Login and menu functions
Have added a menu which is not yet in use and need to be fixed for our purpose. Also added a login function which works by taking username and password, checking it with the credentials in the database and retrieving the connected users info.

The code is not perfect and can be cleaned and fixed for our purpose.