Work in progress: Artist methods in client
Added the ArtistById() method and started working on ConnectUserToArtist() method in client. Also made some changes to UserLogin() and fixed so that the main menu starts from that method, and that it is possible to log out from the program. Also added models and changed so that loggedInUser is an object of User.